refactor: convert scripts directory to typescript
